refactor(sanity): move sanityFetch to dedicated module and update imports
Centralize the `sanityFetch` function in a dedicated module (`sanity/lib/sanityFetch.ts`) to improve code organization and maintainability. Updated all imports across the codebase to use the new location. Removed the redundant `live.ts` file and simplified the `client.ts` file by removing the inline `sanityFetch` implementation.
