+++ sanity/client.ts
@@ -1,4 +1,4 @@
-import { createClient, QueryParams } from 'next-sanity';
+import { createClient } from 'next-sanity';
 
 export const client = createClient({
     projectId: 'p5q8f9ac',
@@ -7,21 +7,3 @@ export const client = createClient({
     token: process.env.NEXT_PUBLIC_SANITY_TOKEN,
     useCdn: true,
 });
-export async function sanityFetch<const QueryString extends string>({
-    query,
-    params = {},
-    revalidate = 60, // default revalidation time in seconds
-    tags = [],
-  }: {
-    query: QueryString
-    params?: QueryParams
-    revalidate?: number | false
-    tags?: string[]
-  }) {
-    return client.fetch(query, params, {
-      next: {
-        revalidate: tags.length ? false : revalidate, // for simple, time-based revalidation
-        tags, // for tag-based revalidation
-      },
-    })
-  }
